conclui fase de rotas no plano
