+++ docs/plano-refatoracao-backend.md
@@ -90,23 +90,26 @@ Este documento serve como guia e controle de progresso para a refatoração arqu
 ### 🚀 Fase 1: Reestruturação de Rotas (Semana 1-2)
 
 #### ✅ Tarefas Concluídas
-- [ ] Análise da arquitetura atual
-- [ ] Definição do plano de refatoração
-- [ ] Criação deste documento de acompanhamento
+- [x] Análise da arquitetura atual
+- [x] Definição do plano de refatoração
+- [x] Criação deste documento de acompanhamento
+- [x] Criar estrutura de pastas para nova arquitetura
+- [x] Separar rotas de autenticação (`auth.routes.ts`)
+- [x] Criar arquivo principal de rotas (`routes/index.ts`)
+- [x] Atualizar arquivo principal para usar nova estrutura de módulos
+- [x] Separar rotas de inventário (`inventory.routes.ts`)
+
+- [x] Separar rotas de produtos (`product.routes.ts`)
+- [x] Separar rotas de usuários (`user.routes.ts`)
+- [x] Separar rotas de relatórios (`report.routes.ts`)
+
+- [x] Testes de regressão
 
 #### 🔄 Em Andamento
 - [ ] Nenhuma tarefa em andamento
 
 #### ⏳ Pendentes
-- [ ] Criar estrutura de pastas para nova arquitetura
-- [ ] Separar rotas de autenticação (`auth.routes.ts`)
-- [ ] Separar rotas de inventário (`inventory.routes.ts`)
-- [ ] Separar rotas de produtos (`product.routes.ts`)
-- [ ] Separar rotas de usuários (`user.routes.ts`)
-- [ ] Separar rotas de relatórios (`report.routes.ts`)
-- [ ] Criar arquivo principal de rotas (`index.routes.ts`)
-- [ ] Atualizar `routes.ts` para usar nova estrutura
-- [ ] Testes de regressão
+- [ ] Nenhuma tarefa pendente
 
 ### 🏗️ Fase 2: Implementação de Serviços (Semana 3-4)
 
@@ -200,16 +203,16 @@ Este documento serve como guia e controle de progresso para a refatoração arqu
 ## 📊 Métricas de Progresso
 
 ### 📈 Status Geral
-- **Progresso Total**: 0% (0/6 fases concluídas)
-- **Fase Atual**: Fase 1 - Reestruturação de Rotas
-- **Tarefas Concluídas**: 3/60 (5%)
-- **Próxima Milestone**: Separação completa das rotas
+  - **Progresso Total**: 17% (1/6 fases concluídas)
+  - **Fase Atual**: Fase 2 - Implementação de Serviços
+  - **Tarefas Concluídas**: 12/60 (20%)
+  - **Próxima Milestone**: Implementação de Serviços
 
 ### 📋 Checklist por Fase
 
 | Fase | Descrição | Status | Progresso |
 |------|-----------|--------|----------|
-| 1 | Reestruturação de Rotas | 🔄 Em Andamento | 3/10 (30%) |
+| 1 | Reestruturação de Rotas | ✅ Concluída | 12/12 (100%) |
 | 2 | Implementação de Serviços | ⏳ Pendente | 0/8 (0%) |
 | 3 | Padrão Repository | ⏳ Pendente | 0/8 (0%) |
 | 4 | Middlewares e Validação | ⏳ Pendente | 0/7 (0%) |
@@ -272,7 +275,7 @@ Este documento deve ser atualizado:
 
 ---
 
-**Última atualização**: Janeiro 2025  
-**Próxima revisão**: Início da Fase 1  
-**Responsável**: Equipe de Desenvolvimento  
-**Status**: 🚀 Iniciado
+**Última atualização**: Agosto 2025
+**Próxima revisão**: Início da Fase 2
+**Responsável**: Equipe de Desenvolvimento
+**Status**: 🚀 Em andamento
