+++ server/routes/auth.routes.ts
@@ -1,0 +1,91 @@
+import type { Express } from "express";
+import { getStorage } from "../db";
+import { isAuthenticated, verifyPassword } from "../auth";
+import { loginSchema, registerSchema } from "@shared/schema";
+
+export function registerAuthRoutes(app: Express) {
+  app.post("/api/auth/login", async (req, res) => {
+    try {
+      const { username, password } = loginSchema.parse(req.body);
+
+      const storage = await getStorage();
+      const user = await storage.getUserByUsername(username);
+      if (!user || !user.isActive) {
+        const message =
+          user && !user.isActive
+            ? "Usuário desativado"
+            : "Usuário não encontrado";
+        return res.status(401).json({ message });
+      }
+
+      const isValidPassword = await verifyPassword(password, user.password);
+      if (!isValidPassword) {
+        return res.status(401).json({ message: "Credenciais inválidas" });
+      }
+
+      const session = req.session as any;
+      session.userId = user.id;
+
+      const { password: _, ...userWithoutPassword } = user;
+      res.json({ user: userWithoutPassword });
+    } catch (error) {
+      console.error("Error during login:", error);
+      res.status(500).json({ message: "Erro interno do servidor" });
+    }
+  });
+
+  app.post("/api/auth/register", async (req, res) => {
+    try {
+      const userData = registerSchema.parse(req.body);
+
+      const storage = await getStorage();
+      const existingUser = await storage.getUserByUsername(userData.username);
+      if (existingUser) {
+        return res.status(400).json({ message: "Nome de usuário já existe" });
+      }
+
+      const existingEmail = await storage.getUserByEmail(userData.email);
+      if (existingEmail) {
+        return res.status(400).json({ message: "Email já cadastrado" });
+      }
+
+      const { confirmPassword, ...userDataWithoutConfirm } = userData;
+
+      const newUser = await storage.createUser({
+        ...userDataWithoutConfirm,
+        role: "user",
+        isActive: true,
+      });
+
+      const session = req.session as any;
+      session.userId = newUser.id;
+
+      const { password: _, ...userWithoutPassword } = newUser;
+      res.status(201).json({ user: userWithoutPassword });
+    } catch (error) {
+      console.error("Error during registration:", error);
+      res.status(500).json({ message: "Erro interno do servidor" });
+    }
+  });
+
+  app.post("/api/auth/logout", (req, res) => {
+    const session = req.session as any;
+    session.destroy((err: any) => {
+      if (err) {
+        return res.status(500).json({ message: "Erro ao fazer logout" });
+      }
+      res.json({ message: "Logout realizado com sucesso" });
+    });
+  });
+
+  app.get("/api/auth/user", isAuthenticated, async (req: any, res) => {
+    try {
+      const { password: _, ...userWithoutPassword } = req.user;
+      res.json(userWithoutPassword);
+    } catch (error) {
+      console.error("Error fetching user:", error);
+      res.status(500).json({ message: "Failed to fetch user" });
+    }
+  });
+}
+
